@@ -19,8 +19,8 @@
 		"@faker-js/faker": "^8.4.1",
 		"@playwright/test": "^1.46.1",
 		"@sveltejs/adapter-auto": "^3.2.4",
-		"@sveltejs/kit": "^2.8.3",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@sveltejs/kit": "^2.12.0",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/tcp-port-used": "^1.0.4",
 		"@typescript-eslint/eslint-plugin": "^7.18.0",
 		"@typescript-eslint/parser": "^7.18.0",
@@ -28,14 +28,14 @@
 		"concurrently": "^8.2.2",
 		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.43.0",
+		"eslint-plugin-svelte": "^2.45.1",
 		"postcss": "^8.4.41",
 		"postgres": "^3.4.4",
 		"prettier": "^3.3.3",
 		"prettier-plugin-svelte": "^3.2.6",
 		"supabase": "^2.54.11",
-		"svelte": "^4.2.19",
-		"svelte-check": "^3.8.6",
+		"svelte": "^5.0.0",
+		"svelte-check": "^4.0.0",
 		"tailwindcss": "^3.4.10",
 		"tcp-port-used": "^1.0.2",
 		"tslib": "^2.7.0",
