@@ -1,8 +1,13 @@
 <script lang="ts">
-	let className = '';
-	export { className as class };
+	interface Props {
+		class?: string;
+		children?: import('svelte').Snippet;
+	}
+
+	let { class: className = '', children }: Props = $props();
+	
 </script>
 
 <div class="alert {className} shadow-lg rounded-none">
-	<slot />
+	{@render children?.()}
 </div>
