@@ -1,5 +1,14 @@
+<script lang="ts">
+	interface Props {
+		children?: import('svelte').Snippet;
+	}
+
+	let { children }: Props = $props();
+</script>
+
 <svelte:head>
-	<title>User Profile</title>
+	<title>Supabase by example</title>
+	<meta name="description" content="Supabase by example showcasing it's auth features.">
 </svelte:head>
 
 <main data-theme="winter">
@@ -7,7 +16,7 @@
 		<div class="navbar border-b border-gray-300 py-8 px-4">
 			<div class="flex-1">
 				<h1 class="font-semibold">
-					<a href="/">User Profile</a>
+					<a href="/">Supabase by example</a>
 				</h1>
 			</div>
 			<div class="flex-none space-x-10">
@@ -18,7 +27,7 @@
 			</div>
 		</div>
 		<div class="grid place-items-center my-20 mx-2">
-			<slot />
+			{@render children?.()}
 		</div>
 	</div>
 </main>
