@@ -1,5 +1,5 @@
 <script lang="ts">
-	export let data;
+	let { data } = $props();
 </script>
 
 <div class="card w-4/12 bg-base-100 shadow-xl">
