@@ -3,7 +3,7 @@
 	import Alert from '$lib/Alert.svelte';
 	import InputErrorMessage from '$lib/InputErrorMessage.svelte';
 
-	export let form;
+	let { form } = $props();
 </script>
 
 <div
