@@ -4,8 +4,12 @@
 	import Alert from '$lib/Alert.svelte';
 	import { enhance } from '$app/forms';
 
-	export let data: PageData;
-	export let form: ActionData;
+	interface Props {
+		data: PageData;
+		form: ActionData;
+	}
+
+	let { data, form }: Props = $props();
 </script>
 
 <div
