@@ -4,7 +4,11 @@
 	import InputErrorMessage from '$lib/InputErrorMessage.svelte';
 	import type { ActionData } from './$types';
 
-	export let form: ActionData;
+	interface Props {
+		form: ActionData;
+	}
+
+	let { form }: Props = $props();
 </script>
 
 <div
