@@ -1,13 +1,13 @@
 <script lang="ts">
-	export let data;
+	let { data } = $props();
 </script>
 
 <div class="min-w-screen min-h-screen flex items-center justify-center px-5 py-5">
 	<div class="rounded-lg shadow-xl bg-gray-900 text-white">
 		<div class="border-b border-gray-800 px-8 py-3">
-			<div class="inline-block w-3 h-3 mr-2 rounded-full bg-red-500" />
-			<div class="inline-block w-3 h-3 mr-2 rounded-full bg-yellow-300" />
-			<div class="inline-block w-3 h-3 mr-2 rounded-full bg-green-400" />
+			<div class="inline-block w-3 h-3 mr-2 rounded-full bg-red-500"></div>
+			<div class="inline-block w-3 h-3 mr-2 rounded-full bg-yellow-300"></div>
+			<div class="inline-block w-3 h-3 mr-2 rounded-full bg-green-400"></div>
 		</div>
 		<div class="px-8 py-6">
 			<p>
